refactor the nasty details of how to use the event loop
We'd *really* prefer not to use the global event loop. It means we have
to do an atexit.register() to close it later, which could conflict with
the user's own cleanup. Also on Windows we'll need to make sure we use
a ProactorEventLoop, and we won't want to set that globally. Do my best
to contain all this crap in a separate function.