get rid of OrThen, and add the check command arg
I think OrThen is going to be used almost entirely for suppressing
errors, in the form `.orthen('true')`. Better to get rid of that method
and support suppressing errors directly. Allow commands (in addition to
run methods) to take the "check" keywork argument. On commands, this
forces the return status of the program to be 0.

Defining this first command flag raised an issue with how our API is
going to be shaped. Consider this pipeline:

    a.pipe(b).pipe(c, check=False).run()

That translates into an object hierarchy that looks like this:

    Pipe
    /  \
  Pipe  c
  /  \
 a    b

So we have to decide whether the "check" argument should apply just to
the command c, or to the whole Pipe at the top. I think it should apply
only to c, for a couple reasons:
1) We need *some* syntax for giving arguments to just one command, and
   it's hard to imagine what else it could be.
2) We don't want the caller to have to think about the object graph. The
   method chain looks flat, like a shell pipeline, and we want to
   preserve that intuition. Applying arguments to compound expressions
   would force the caller to think about which object ends up on top.