replace InputWriter and OutputReader with IOContext
When we add input/stdin/stdout/stderr parameters to individual commands,
we're going to need to share the logic for those with the run methods.
This includes new logic like opening filepaths and interpreting
STDOUT/STDERR.

This diff implements the full set of IO flags in IOContext, and calls
into that from the run method. Upcomming diffs will add more callers to
this at the expression level.