default to universal newlines mode in Python 2
This makes a difference in our tests on Windows, where we start seeing a
bunch of '\r' characters in Python 2.7. We need to explicitly set the
'rU' mode, and we also need avoid comparing for identity with the bytes
type, because str and bytes are in fact the same in Python 2.