drop Python 2.6 testing on Appveyor
Instead of importing a ton of AppVeyor scripts from
https://github.com/ogrisel/python-appveyor-demo to handle installing
Python 2.6, just drop it. I did a test run by hand on my own machine,
and all the tests seem to pass, so continuing to test 2.7 is probably
good enough. We can revisit this if we ever get a Windows-Python-2.6
user :p