change STDOUT and STDERR to join simultaneous redirects
Previously if you did something like:

    cmd('foo').run(stdout='myfile', stderr=STDOUT)

we were interpreting that to mean that stderr should go to the *old*
stdout, rather than to myfile. That interpretation becomes especially
confusing in this case:

    cmd('bar').read(stderr=STDOUT)

Here despite appearances, stderr was not captured along with stdout. In
addition to being confusing, this was also different from how the
standard library behaves.

Instead, now we interpret stderr=STDOUT to mean stderr joins the *new*
value of stdout. The only tricky bit is that now it's not immediately
clear what setting stdout=STDERR and stderr=STDOUT at the same time
should mean. But there are only two options:

1) Make it an error.
2) Swap the old values.

Going with (2) means you can still accomplish all the funny swaps you
were able to do with the old semantics. And in the implementation, it
turns out it isn't much of a special case at all. Nice!