get rid of STRING and BYTES in favor of PIPE
I think it will be fairly common for people to want to `read()` stdout
as bytes. Requiring callers to `from duct import BYTES` feels
unnecessarily awkward. Setting `decode=False` is much easier, and it
avoids the confusing state where one output is decoded and the other
isn't.

There's a chance that `PIPE` itself will feel too close to `pipe()` for
comfort, given that it means something very different. I might want to
bikeshed that some more in the future. (Maybe simpler flags like
`capture_stdout` and `capture_stderr`? Though it would be a little weird
to have those and also `stdout`/stderr`.) But in any case `PIPE` is the
Python convention, and it's a reasonable first improvement.

While we're at it, rename `trim` to `sh_trim`. That's more suggestive of
what it's actually doing, especially with an eye towards languages (like
Rust) that use `trim` to mean what `strip` means in Python.

Closes https://github.com/oconnor663/duct/issues/32.