bump version to 0.3.0
This version represents a complete rewrite in the duct.rs style. Keyword
arguments are gone, replaced by method calls. The `subshell` method is
gone, since now every method call is sort of an implicit subshell. It
actually works better this way, because the order of operations is
clearer, especially when you redirect stdout and stderr at the same
time.