get rid of stdin_reader/stdout_writer/stderr_writer
These were something of a footgun, in that if you ran the same
expression twice using the same reader/writer it'd be likely to produce
confusing results. They also don't translate well into Rust, because
often the caller will want to get their hands on a Writer after the
command is done. It's unlikely that this feature would be used enough to
justify the maintenance burden (or even the burden of having users
understand what it means when they skim the docs), and it's relatively
easy for users to emulate with pipes.