remove the ReaderHandle.close method
It was more useful when it killed the child process, but now that it
just closes the read pipe, I think it's more confusing than useful. It
might be nice to close the pipe aggressively in error cases, but you're
still leaking the child and relying on finalization to clean up most of
your resources. ReaderHandle still inherits from IOBase, so it still has
the default no-op .close() method, and it's still technically compatible
with `with` statements. (They just don't do anything now.)